+++ example/lib/custom_dialog_manager.dart
@@ -3,17 +3,17 @@ import 'package:flutter_easy_dialogs/flutter_easy_dialogs.dart';
 
 int? _customDialogId;
 
-class CustomDialogAgent
-    extends EasyDialogAgentBase<CustomAgentShowParams, AgentHideParams?> {
-  CustomDialogAgent({required super.overlayController});
+class CustomDialogManager extends EasyDialogManagerBase<CustomManagerShowParams,
+    ManagerHideParamsBase?> {
+  CustomDialogManager({required super.overlayController});
 
   @override
-  Future<void> hide({AgentHideParams? params}) async {
+  Future<void> hide({ManagerHideParamsBase? params}) async {
     super.overlayController.removeCustomDialog(_customDialogId!);
   }
 
   @override
-  Future<void> show({required CustomAgentShowParams params}) async {
+  Future<void> show({required CustomManagerShowParams params}) async {
     _customDialogId = super.overlayController.insertCustomDialog(
           Align(
             alignment: Alignment.bottomCenter,
@@ -28,10 +28,10 @@ class CustomDialogAgent
   }
 }
 
-class CustomAgentShowParams extends AgentShowParams {
+class CustomManagerShowParams extends ManagerShowParamsBase {
   final Color color;
 
-  const CustomAgentShowParams({
+  const CustomManagerShowParams({
     required super.theme,
     required super.content,
     required this.color,