+++ example/lib/easy_router.dart
@@ -1,4 +1,4 @@
-import 'package:example/screens/custom_agent_screen.dart';
+import 'package:example/screens/custom_agent_manager.dart';
 import 'package:example/screens/home_screen.dart';
 import 'package:example/screens/modal_banner_screen.dart';
 import 'package:example/screens/positioned_dialogs_screen.dart';