+++ lib/flutter_easy_dialogs.dart
@@ -1,10 +1,16 @@
 library flutter_easy_dialogs;
 
-export 'src/core/agents/dialog_agent_base.dart';
 export 'src/core/animations/animations.dart' hide EasyAnimationSettings;
 export 'src/core/dialogs/dialogs.dart';
 export 'src/core/dialogs/easy_modal_banner/easy_modal_banner_theme_data.dart';
 export 'src/core/dismissibles/dismissibles.dart';
 export 'src/core/dismissibles/types/easy_positioned_dismissible_type.dart';
 export 'src/core/flutter_easy_dialogs/flutter_easy_dialogs_theme.dart';
+export 'src/core/managers/managers.dart'
+    show
+        EasyDialogManagerBase,
+        ManagerHideParamsBase,
+        ManagerShowParamsBase,
+        BlockAndroidBackButtonMixin,
+        SingleAutoDisposalControllerMixin;
 export 'src/core/overlay/overlay.dart' hide EasyOverlayAppEntry, EasyOverlay;