+++ lib/src/core/managers/managers.dart
@@ -1,0 +1,9 @@
+export 'easy_dialog_manager_base.dart';
+export 'full_screen_dialog_manager/full_screen_dialog_hide_params.dart';
+export 'full_screen_dialog_manager/full_screen_dialog_manager.dart';
+export 'full_screen_dialog_manager/full_screen_dialog_show_params.dart';
+export 'mixin/block_android_back_button_mixin.dart';
+export 'mixin/single_auto_disposal_controller_mixin.dart';
+export 'positioned_dialog_manager/positioned_dialog_hide_params.dart';
+export 'positioned_dialog_manager/positioned_dialog_manager.dart';
+export 'positioned_dialog_manager/positioned_dialog_show_params.dart';